Added a development file
This file contains code that shows the syntax for using mysql.connector as well as some ideas for the functions that we are requiring for an admin. Still working on the general user. This should speed things up, so feel free to test the code yourselves since I got it to all work so far.